Se corrige crear pagare

+++ src/main/java/com/imss/sivimss/formatopagare/beans/PagareServicio.java
@@ -34,7 +34,7 @@ public class PagareServicio {
 	
 	private Integer id;
 	private Integer idODS;
-	private Integer importe;
+	private Double importe;
 	private Double redito;
 	private Integer idUsuarioAlta;
 	private Integer idUsuarioModifica;
